Reuse date formatter in autoform
